% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_functions.R
\docType{data}
\name{acc_data_train}
\alias{acc_data_train}
\title{Data about the 400 accounts in the training set}
\description{
For each individual (id), we have the group of the account (group) bot (group = 2) and genuine (group = 1). Additionally for each individual, we have the number of accounts which the individual follows (friends\_count), the number of accounts which follow the individual (followers\_count), and the date when the account was created (created\_at).
}
\author{
Anthony Weishampel \email{acweisha@ncsu.edu}
}
\keyword{Data}
\keyword{Functional}
