% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bivariate_smooth}
\alias{bivariate_smooth}
\title{Estimate smooth covariance function}
\usage{
bivariate_smooth(K_b, k = 10, bs0 = "cr")
}
\arguments{
\item{K_b}{NJ x D matrix of observations}

\item{k}{number of basis functions in the}

\item{bs0}{The basis functions used when smoothing the covariance functions}
}
\value{
Smooth covariance functions
}
\description{
Estimate smooth covariance function
}
