% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{d_logit}
\alias{d_logit}
\title{function to return the derivative of the logit}
\usage{
d_logit(x)
}
\arguments{
\item{x}{value to input}
}
\value{
Returns the derivative of the logit of x
}
\description{
function to return the derivative of the logit
}
