% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_eigenfunctions}
\alias{estimate_eigenfunctions}
\title{Function to estimate eigenfunctions}
\usage{
estimate_eigenfunctions(K_b, pve = 0.98, fix_num_of_functions = NA)
}
\arguments{
\item{K_b}{estimate of a covariance matrix}

\item{pve}{Proportion of variance explained}

\item{fix_num_of_functions}{set the number of eigenfunctions to be returned}
}
\value{
The eigenvalues and eigenvectors from
}
\description{
Function to estimate eigenfunctions
}
