% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_eigenfunctions2}
\alias{estimate_eigenfunctions2}
\title{Function to estimate eigenfunctions when there are two covariance functions}
\usage{
estimate_eigenfunctions2(
  K_b,
  K_w,
  pve1 = 0.95,
  pve2 = 0.95,
  fix_num_of_functions1 = NA,
  fix_num_of_functions2 = NA
)
}
\arguments{
\item{K_b}{: Covariance matrix for the first level component}

\item{K_w}{: Covariance matrix for the second level component}

\item{pve1}{: Proportion of variance explained for the first component}

\item{pve2}{: Proportion of variance explained for the second component}

\item{fix_num_of_functions1}{if you do not want to used PVE you can fix the number of eigenfunctions to output}

\item{fix_num_of_functions2}{if you do not want to used PVE you can fix the number of eigenfunctions to output}
}
\value{
Eigenfunctions and eigenvalues for both types of covariance functions
}
\description{
Function to estimate eigenfunctions when there are two covariance functions
}
