% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_ajs_model}
\alias{fit_ajs_model}
\title{function used to train the gAR models}
\usage{
fit_ajs_model(l, j_lags, s_mat_train, aic = F, classes, static_train = NA)
}
\arguments{
\item{l}{the group which is currently being used to fit gAR models}

\item{j_lags}{the lag used in the gAR}

\item{s_mat_train}{estimated latent state data used to fit the gAR model}

\item{aic}{T/F to return the AIC of the model}

\item{classes}{vector of the classes}

\item{static_train}{dataframe containing the information of covariates which one would like to include in the gAR model}
}
\value{
Fitted gAR model
}
\description{
function used to train the gAR models
}
