% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func.cur}
\alias{func.cur}
\title{Function used in the estimation of the covariance functions
#' Function used in covariance estimation used for}
\usage{
func.cur(i.cur, X_dat, N)
}
\arguments{
\item{i.cur}{cur index}

\item{X_dat}{Binary Valued Functional Data}

\item{N}{number of individuals}
}
\description{
Function used in the estimation of the covariance functions
#' Function used in covariance estimation used for
}
