% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_binary_fns}
\alias{generate_binary_fns}
\title{Function used to generate the binary valued functional data from latent curves}
\usage{
generate_binary_fns(X)
}
\arguments{
\item{X}{is the latent functional data}
}
\value{
Binary valued functional data determined by X
}
\description{
Function used to generate the binary valued functional data from latent curves
}
