% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_data_with_S}
\alias{generate_data_with_S}
\title{Functions used to generate latent state data in the various scenarios in accompanying paper}
\usage{
generate_data_with_S(
  scenario = 3,
  Curves_binary,
  N,
  J,
  Ys,
  alpha1 = 2/3,
  alpha2 = 1/4
)
}
\arguments{
\item{scenario}{Which scenario to generate S data}

\item{Curves_binary}{The binary-valued functional data}

\item{N}{Total number of subjects to generate data for}

\item{J}{Number of realizations}

\item{Ys}{Group assignments for the N individuals}

\item{alpha1}{probability for latent state i ngroup 1}

\item{alpha2}{probability for latent state in group 2}
}
\value{
Binary or latent curves with latent states
}
\description{
Functions used to generate latent state data in the various scenarios in accompanying paper
}
