% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_latent_process}
\alias{generate_latent_process}
\title{Function used to generate the latent curves for simulation scenarios where I define various scenarios, inspired by Delaigle and Hall(2012)}
\usage{
generate_latent_process(theta, mu_coeff, tt, n = 20, return_scores_too = F)
}
\arguments{
\item{theta}{Define latent process X_i using Fourier basis}

\item{mu_coeff}{vector of mean coefficients}

\item{tt}{grid of points which the functions are observed}

\item{n}{number of individuals to generate data for}

\item{return_scores_too}{Indicator to determine whether or not to return the coefficients too}
}
\value{
latent process observed on points defined by grid
}
\description{
Function used to generate the latent curves for simulation scenarios where I define various scenarios, inspired by Delaigle and Hall(2012)
}
