% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_length_pve}
\alias{get_length_pve}
\title{Function: Get number of functions based on the pve and eigenvalues}
\usage{
get_length_pve(pve, vec, set_max_number = NA)
}
\arguments{
\item{pve}{Value [0,1] to determine number of eigenfunctions}

\item{vec}{vector of eigenvalues}

\item{set_max_number}{If you want a maximum number of values}
}
\value{
The number of eigenfunctions and eigenvalues for KL-approximation
}
\description{
Function: Get number of functions based on the pve and eigenvalues
}
