% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mat_i}
\alias{get_mat_i}
\title{Function used in covariance estimation used for}
\usage{
get_mat_i(i, j1, X_dat, N)
}
\arguments{
\item{i}{cur index of i}

\item{j1}{cur index of j}

\item{X_dat}{Binary Valued Functional Data}

\item{N}{number of individuals}
}
\description{
Function used in covariance estimation used for
}
