% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pdf_den2}
\alias{get_pdf_den2}
\title{Function: To get the density values of new scores for an individual in a given class}
\usage{
get_pdf_den2(densities, scores, classes, i)
}
\arguments{
\item{densities}{List of densities of scores within each group}

\item{scores}{matrix of scores for the N individuals}

\item{classes}{matrix of classes (these are not the true classes but the value to be evaluated)}

\item{i}{which individual to evaluate}
}
\value{
Density values of the new scores for individual i
}
\description{
Function: To get the density values of new scores for an individual in a given class
}
