% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_true_alpha}
\alias{get_true_alpha}
\title{function to get true properties of S for the simulation study}
\usage{
get_true_alpha(scenario = 1, J)
}
\arguments{
\item{scenario}{Scenario used to simulation the data}

\item{J}{number of realizations of functional data}
}
\description{
function to get true properties of S for the simulation study
}
