% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_true_bases}
\alias{get_true_bases}
\title{function to get basis for simulation study}
\usage{
get_true_bases(scenario = 1, D)
}
\arguments{
\item{scenario}{Scenario used to simulate the data}

\item{D}{number of points in the grid}
}
\description{
function to get basis for simulation study
}
