% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_functions.R
\name{gMFPCA}
\alias{gMFPCA}
\title{Function for fitting the gmFPCA}
\usage{
gMFPCA(
  X_dat_m,
  Ys,
  J,
  N,
  covariates = NA,
  gAR = F,
  pve1 = 0.95,
  pve2 = 0.95,
  Kb = 5,
  q = 3,
  approximation = "linear",
  gar_covariates = NA,
  bs0 = "cr"
)
}
\arguments{
\item{X_dat_m}{N*J x m matrix of binary data}

\item{Ys}{N long vector of responses}

\item{J}{Number of realizations per individuals}

\item{N}{number of individuals}

\item{covariates}{N x Q Data frame of covariates}

\item{gAR}{variable to include generalized autoregressive structure}

\item{pve1}{Proportion of variation explained in the first level eigenfunction}

\item{pve2}{Proportion of variation explained in the first level eigenfunction}

\item{Kb}{number of basis functions in smoothing covariance functions}

\item{q}{Lag in the generalized Auto regressive model}

\item{approximation}{choice of which approximation to use}

\item{gar_covariates}{covariates to include in the gAR model}

\item{bs0}{type of basis functions when smoothing covariance functions}
}
\value{
list of information required to build the gmFPCA model and predict new groups
}
\description{
Function for fitting the gmFPCA
}
