% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_functions.R
\name{gsFPCA}
\alias{gsFPCA}
\title{Function for modeling single-level fpca}
\usage{
gsFPCA(
  X_dat_s,
  Ys,
  covariates = NA,
  pve = 0.95,
  Kb = 10,
  num_knots = 10,
  bs0 = "cr"
)
}
\arguments{
\item{X_dat_s}{N x m matrix of binary data}

\item{Ys}{N long vector of groups}

\item{covariates}{N x Q data frame of covariates}

\item{pve}{Proportion of variation explained}

\item{Kb}{number of basis functions used when estimating latent trajectories}

\item{num_knots}{number of knots in basis functions}

\item{bs0}{type of basis functions used when estimating latent trajectories}
}
\value{
list of information required to build the model and predict new groups
}
\description{
Function for modeling single-level fpca
}
