% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_functions.R
\name{gsFPCA_predict}
\alias{gsFPCA_predict}
\title{Function for predicting new groups of based on single realizations of binary-valued functional data}
\usage{
gsFPCA_predict(gsFPCA.model, X_dat_s_new, covariates_new = NA)
}
\arguments{
\item{gsFPCA.model}{trained gsFPCA model}

\item{X_dat_s_new}{N_new x m matrix of binary data}

\item{covariates_new}{N_new x Q data frame of covariates}
}
\value{
Predicted new groups for the N_new users
}
\description{
Function for predicting new groups of based on single realizations of binary-valued functional data
}
