% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{invlogit}
\alias{invlogit}
\title{Function to return the inverse of the logit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{value to input}
}
\value{
Returns the invlogit of x
}
\description{
Function to return the inverse of the logit
}
