% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{k_v_func}
\alias{k_v_func}
\title{Function used to estimate the covariance functions in the multilevel scenario}
\usage{
k_v_func(i.cur, X_dat, N)
}
\arguments{
\item{i.cur}{cur index}

\item{X_dat}{Current indexed binary-valued functional data}

\item{N}{number of individuals}
}
\value{
estimate averaged over j
}
\description{
Function used to estimate the covariance functions in the multilevel scenario
}
