% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{K_w_func_over_j}
\alias{K_w_func_over_j}
\title{Function used to estimate the covariance functions in the multilevel scenario}
\usage{
K_w_func_over_j(i, j1, X_dat.cur)
}
\arguments{
\item{i}{cur index}

\item{j1}{number of individuals}

\item{X_dat.cur}{Current indexed binary-valued functional data}
}
\value{
estimate averaged over j
}
\description{
Function used to estimate the covariance functions in the multilevel scenario
}
