% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logit}
\alias{logit}
\title{Function to return the logit of x}
\usage{
logit(x)
}
\arguments{
\item{x}{value to input}
}
\value{
Returns the logit of x
}
\description{
Function to return the logit of x
}
