% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_pos_semi_def}
\alias{make_pos_semi_def}
\title{Function that makes a matrix positive semi definite}
\usage{
make_pos_semi_def(x)
}
\arguments{
\item{x}{matrix to make semi-positive definite}
}
\value{
semi-positive definite of x
}
\description{
Function that makes a matrix positive semi definite
}
