% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multilevel_linear_fpca}
\alias{multilevel_linear_fpca}
\title{Estimate mean function, eigenfunctions, and eigenvalues  for the multilevel model using the Linear approximation}
\usage{
multilevel_linear_fpca(
  X_dat,
  J,
  pve1 = 0.95,
  pve2 = 0.95,
  fix_number_of_functions = NA,
  k = 5,
  bs0 = "cr"
)
}
\arguments{
\item{X_dat}{NJ x D matrix of observations}

\item{J}{Number of realizations per individual}

\item{pve1}{Proportion of variance explained for the first component}

\item{pve2}{Proportion of variance explained for the second component}

\item{fix_number_of_functions}{if you do not want to used PVE you can fix the number of eigenfunctions to output}

\item{k}{number of basis functions in the}

\item{bs0}{The basis functions used when smoothing the covariance functions}
}
\value{
Mean function, eigenfunctions and eigenvalues for both types of covariance functions
}
\description{
Estimate mean function, eigenfunctions, and eigenvalues  for the multilevel model using the Linear approximation
}
