% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nb_updated}
\alias{nb_updated}
\title{Function: Estimate predicted values for a given value of scaling parameter h}
\usage{
nb_updated(
  scores,
  classes,
  prior_g,
  scores_test,
  s_mat_hat_train,
  s_mat_hat_test,
  alpha_js = NA,
  h = 1.06,
  P_max = 3,
  static_train = NA,
  static_test = NA
)
}
\arguments{
\item{scores}{N x K matrix of scores in the training set}

\item{classes}{Group labels vector of length N}

\item{prior_g}{vector of prior probability of being in each group, sums up to 1}

\item{scores_test}{N_test x K matrix of scores in the testing set}

\item{s_mat_hat_train}{matrix of estimated S latent state values training set}

\item{s_mat_hat_test}{matrix of estimated S latent state values testing set}

\item{alpha_js}{Estimated probability of latent states across the realizations}

\item{h}{multiplier for kernel based density function}

\item{P_max}{Lag in the gAR models}

\item{static_train}{Covariates for the classifier in the training set accounts}

\item{static_test}{Covariates for the classifier in the testing set accounts}
}
\value{
Predicted groups for the individuals
}
\description{
Function: Estimate predicted values for a given value of scaling parameter h
}
