% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nb_updated_grid}
\alias{nb_updated_grid}
\title{Function: Grid search to estimate predicted values and estimate the values of h}
\usage{
nb_updated_grid(
  scores,
  s_mat_hat_train,
  classes,
  prior_g,
  scores_test,
  s_mat_hat_test,
  alpha_js = NA,
  min.h = 0.4,
  max.h = 1.6,
  CV = 5,
  n_grid = 5,
  return_h = F,
  return_prob = F,
  P_max = 3,
  static_train = NA,
  static_test = NA
)
}
\arguments{
\item{scores}{N x K matrix of 1st level scores in the training set}

\item{s_mat_hat_train}{Latent state matrix of estimated values for the training set accounts}

\item{classes}{Group labels vector of length N}

\item{prior_g}{vector of prior probability of being in each group, sums up to 1}

\item{scores_test}{N_test x K matrix of scores in the testing set}

\item{s_mat_hat_test}{Latent state matrix of estimated values for the testing set accounts}

\item{alpha_js}{Estimated probability of latent states across the realizations}

\item{min.h}{min possible value for the multiplier}

\item{max.h}{maximum possible value for the multiplier}

\item{CV}{Number of folds for cross validation}

\item{n_grid}{number of values between min.h and max.h to search over}

\item{return_h}{T/F to return the value of the multiplier}

\item{return_prob}{T/F to return group Bayes classifier probability for each individual in testing set}

\item{P_max}{Lag of the gAR models}

\item{static_train}{Covariates for the classifier in the training set accounts}

\item{static_test}{Covariates for the classifier in the testing set accounts}
}
\value{
Predicted group for each individual
}
\description{
Function: Grid search to estimate predicted values and estimate the values of h
}
