% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nb_updated_grid_scores_only}
\alias{nb_updated_grid_scores_only}
\title{Function: Grid search to estimate predicted values and estimate the values of h}
\usage{
nb_updated_grid_scores_only(
  scores,
  classes,
  prior_g,
  scores_test,
  min.h = 0.01,
  max.h = 1.5,
  CV = 10,
  n_grid = 10,
  return_h = F,
  return_prob = F
)
}
\arguments{
\item{scores}{N x K matrix of 1st level scores in the training set}

\item{classes}{Group labels vector of length N}

\item{prior_g}{vector of prior probability of being in each group, sums up to 1}

\item{scores_test}{N_test x K matrix of scores in the testing set}

\item{min.h}{min possible value for the multiplier}

\item{max.h}{maximum possible value for the multiplier}

\item{CV}{Number of folds for cross validation}

\item{n_grid}{number of values between min.h and max.h to search over}

\item{return_h}{T/F to return the value of the multiplier}

\item{return_prob}{T/F to return group Bayes classifier probability for each individual in testing set}
}
\value{
predictions from the grid search
}
\description{
Function: Grid search to estimate predicted values and estimate the values of h
}
