% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_new_ind_group_model}
\alias{predict_new_ind_group_model}
\title{function used to predict the group for new individuals based on the latent states and the fitted gAR models}
\usage{
predict_new_ind_group_model(i, l, p, models_ls, s_mat_test, static_test = NA)
}
\arguments{
\item{i}{index of the current individual}

\item{l}{the group which is currently being used}

\item{p}{the lag of the current models}

\item{models_ls}{Fitted gAR models for the various groups}

\item{s_mat_test}{estimated latent state data used to predict the gAR model}

\item{static_test}{dataframe containing the information of covariates which one would like to include in the gAR model}
}
\value{
Predicted probability of being in group l for the new individuals
}
\description{
function used to predict the group for new individuals based on the latent states and the fitted gAR models
}
