% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regression_bf2}
\alias{regression_bf2}
\title{Function: wrapper function for bayesglm() which outputs expected coefficients for the gaussian responses}
\usage{
regression_bf2(z, dta, lm_structure, eigen_vals1)
}
\arguments{
\item{z}{index the n individual}

\item{dta}{: data frame contain the subject id, mean function, eigenfunctions and observe binary values}

\item{lm_structure}{formula displaying the bayesglm function}

\item{eigen_vals1}{eigen_values for the psi coefficients}
}
\value{
Fitted values from the game function for subject z
}
\description{
Function: wrapper function for bayesglm() which outputs expected coefficients for the gaussian responses
}
