% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regression_g}
\alias{regression_g}
\title{Generalized additive model function gam() fits the observed binary-valued functional data by fitting a GLMM using a set of known basis functions}
\usage{
regression_g(z, Curves, tt, k = 10, method = "REML", bs0 = "cr")
}
\arguments{
\item{z}{index z = 1,...,N}

\item{Curves}{N x D matrix of observed binary series}

\item{tt}{grid of timepoints going from 0 to 1 with D observations}

\item{k}{number of basis functions}

\item{method}{method used to evaluate the gam}

\item{bs0}{type of basis functions used in the gam}
}
\value{
Fitted values from the game function for subject z
}
\description{
Generalized additive model function gam() fits the observed binary-valued functional data by fitting a GLMM using a set of known basis functions
}
