% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_functions.R
\docType{data}
\name{X_dat_m_test}
\alias{X_dat_m_test}
\title{The multilevel binary-valued functional data summarizing 2-weeks of posting data for the 100 accounts in the testing set}
\description{
The multilevel binary-valued functional data summarizing 2-weeks of posting data for the 100 accounts in the testing set
}
\author{
Anthony Weishampel \email{acweisha@ncsu.edu}
}
\keyword{Binary-Valued}
\keyword{Data}
\keyword{Functional}
